fix overflow, fix pixel copy from 8x8 block to bitmap

+++ struct.h
@@ -96,7 +96,7 @@ typedef struct {
     uint8_t Ta;
     uint16_t width;
     uint16_t height;
-    uint16_t pred;
+    int16_t pred;
 } Component;
 
 typedef struct {
@@ -105,7 +105,7 @@ typedef struct {
     uint8_t maxH, maxV;
     uint8_t n_comp;
     Component comp[4];
-    int8_t qt_zz[4][64];
+    int16_t qt_zz[4][64];
     Node *huf[2][2];
     uint8_t **bmp_YCbCr[4];
     uint8_t **bmp_RGB[4];
